+++ src/bindings/yew.rs
@@ -1,6 +1,4 @@
 //! Yew integration module.
-//! The user doesn't need to do anything but to put a style into the class of a
-//! yew component.
 
 use crate::Style;
 use yew::html::Classes;